requirements.txt
python-telegram-bot==21.6
httpx==0.27.2
APScheduler==3.10.4
python-dotenv==1.0.1
aiosqlite==0.20.0
